prov/verbs: Make verbs fi_getinfo rx_attr verification thread-safe
Handle the case where the application has not requested FI_RX_CQ_DATA
in the RX attribute check by removing FI_RX_CQ_DATA in a copy of the
default fi_info. Note, that this fix also removes the check of
domain->cq_data_size as a method of determining if FI_RX_CQ_DATA is
required. It is an error to set domain->cq_data_size and not also
include FI_RX_CQ_DATA.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>