prov/rxm: Allow data FI_PROGRESS_AUTO with FI_ATOMIC capability
Enable the RXM emulated FI_ATOMIC protocol to support one-sided
use cases where the target of the atomics is using data FI_PROGRESS_AUTO,
and is not calling into the API to make progress.

This progress will only be enabled if FI_ATOMIC is requested with data
FI_PROGRESS_AUTO and threading is either FI_THREAD_SAFE or FI_THREAD_UNSPEC,
and leverages the existing CM thread for this progress.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>