prov/verbs: Remove XRC domain file when no longer open by a process
This is a linux specific solution to cleanup the XRC domain file shared
among processes within the same XRC domain. This solution will cleanup
the file under normal operation, but will not remove the file in the event
of a process abort. If an abort occurs and the next process to open the
XRC domain file is not running as the same user, then the file will need
to be manually removed. The implementation attempts to use file descriptor
locking not process, so that a single process can create multiple domains
and the code will work as intended (this requires linux 3.15 and later).

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>