util/memhooks: update brk and implement sbrk intercepts
Previously, the `brk` invalidated a region in cache when the call
resulted in the heap growing, the changes to this function correct
that to occur when the heap shrinks.

Calls to `sbrk` can invalidate MRs on the heap but were not being
intercepted, this introduces an `sbrk` intercept to handle that case.

These changes address the `brk`/`sbrk` failures encountered when using
memhooks with the new mr_cache_evict fabtest (#6029).

Signed-off-by: Peter Gottesman <pgottes@amazon.com>