prov/verbs: Refine XRC INI connection object locking
NOTE: This commit is exprimental and is not operational
      with arkouda/chapel yet.

XRC is intended to be used layered with the RxM provider.
RxM EP open a unique message provider TX CQ; and by definition
an XRC INI connection can only be shared if both EP are
using the same TX CQ. Therefore, INI connections themselves
will not be shared between RxM EP and we can reduce the
duration of holding the ini_lock to protecting INI connection
lockup/creation/destruction.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>